Add new settings provider to popup. Refactor files in providers.
