+++ packages/extension/src/constants.ts
@@ -41,3 +41,4 @@ export const DEVTOOLS_SET_JAVASCSCRIPT_COOKIE =
 export const SET_TAB_TO_READ = 'SET_TAB_TO_READ';
 export const INITIAL_SYNC = 'INITIAL_SYNC';
 export const NEW_COOKIE_DATA = 'NEW_COOKIE_DATA';
+export const SERVICE_WORKER_RELOAD_MESSAGE = 'ServiceWorker::TABS_RELOADED';
