+++ packages/extension/src/view/popup/stateProviders/cookie/useCookie.ts
@@ -13,29 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/**
- * External dependencies.
- */
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-
 /**
  * Internal dependencies.
  */
-import App from './app';
-import { CookieProvider, SettingsProvider } from './stateProviders';
+import useContextSelector from '../../../../utils/useContextSelector';
+import Context, { type CookieStoreContext } from './context';
 
-const root = document.getElementById('root');
+export function useCookie(): CookieStoreContext;
+export function useCookie<T>(selector: (state: CookieStoreContext) => T): T;
 
-if (root) {
-  createRoot(root).render(
-    <SettingsProvider>
-      <CookieProvider>
-        <div className="w-96 min-h-[318px] h-full p-5 mt-4 flex justify-center items-center flex-col">
-          <App />
-        </div>
-      </CookieProvider>
-    </SettingsProvider>
-  );
+/**
+ * Cookie store hook.
+ * @param selector Selector function to partially select state.
+ * @returns selected part of the state
+ */
+export function useCookie<T>(
+  selector: (state: CookieStoreContext) => T | CookieStoreContext = (state) =>
+    state
+) {
+  return useContextSelector(Context, selector);
 }
+
+export default useCookie;
