+++ packages/extension/src/view/popup/stateProviders/settings/context.ts
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * External dependencies.
+ */
+import { createContext } from 'use-context-selector';
+import { noop } from '@ps-analysis-tool/common';
+
+export interface SettingsStoreContext {
+  state: {
+    allowedNumberOfTabs: string | null;
+    isUsingCDP: boolean;
+    settingsChanged: boolean;
+    allowedNumberOfTabsForSettingsDisplay: string | null;
+    isUsingCDPForSettingsDisplay: boolean;
+  };
+  actions: {
+    handleSettingsChange: () => void;
+    setUsingCDP: (newValue: boolean) => void;
+    setSettingsChanged: (newValue: boolean) => void;
+  };
+}
+
+const initialState: SettingsStoreContext = {
+  state: {
+    allowedNumberOfTabs: null,
+    isUsingCDP: false,
+    settingsChanged: false,
+    allowedNumberOfTabsForSettingsDisplay: null,
+    isUsingCDPForSettingsDisplay: false,
+  },
+  actions: {
+    handleSettingsChange: noop,
+    setUsingCDP: noop,
+    setSettingsChanged: noop,
+  },
+};
+
+export default createContext<SettingsStoreContext>(initialState);
