+++ packages/extension/src/view/popup/stateProviders/settings/index.ts
@@ -13,29 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/**
- * External dependencies.
- */
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-
-/**
- * Internal dependencies.
- */
-import App from './app';
-import { CookieProvider, SettingsProvider } from './stateProviders';
-
-const root = document.getElementById('root');
-
-if (root) {
-  createRoot(root).render(
-    <SettingsProvider>
-      <CookieProvider>
-        <div className="w-96 min-h-[318px] h-full p-5 mt-4 flex justify-center items-center flex-col">
-          <App />
-        </div>
-      </CookieProvider>
-    </SettingsProvider>
-  );
-}
+export { default as SettingsProvider } from './settingsProvider';
+export { default as SettingsContext } from './context';
+export { default as useSettings } from './useSettings';
